Finished Strong Authenticated Encryption (v1.5)
Strong Authenticated Encryption is now "complete" in this tool. The nice thing about the algorithm is that the HMAC also serves as the IV and as salt for the key derivation after the exchange.

This guarantees a different key and a different IV for each file for each exchange between two parties.

Next on the agenda are multi-party files.