Added Multiple Recipient Support, Pipes, Etc. (v1.75)
This version adds support for piping in input, the ability to send files to multiple recipients, and the ability to add extensions to encrypted files.

This version will be both future and backwards compatible with version 2.0, since the majority of the work for 2.0 is already completed.

There were also a few major bugs fixed that could've led to the file encoding improperly.

v2.0 will add signatures and contact file extensions, and a hopefully superior CLI.

"Issues":
Currently the sender is always "anonymous" and receives a randomized password, this will be tweaked in version 2.0

This code is not backwards compatible because of a tweak made to the FileProperties struct.