Improve File Format Security.
There was a small "security bug" that was of no use to an attacker, but
could allow an attacker to inconvenience an end user without detection.

It allowed an active attacker to modify the extension count without the
end user detecting it.

This would've led to an output file exporting incorrectly while the tool
would've insisted that everything was fine.

I corrected this by including the extension count in the hmac of the
file.

Other parameters don't need to be included as they are integral to the
decryption process. They would be detected if modified.