Started Migration Away from cppcrypto towards Crypto++ and a few other libraries.
This migration caused a temporary drop in support for the Kuznyechik cipher. It also uncovered a few minor bugs. None really affected security, but I might fix the bug in a later version.

I will add Kuznyechik back in at a later date.

All of the ciphers are now using a Crypto++ backend, which should make it easy to support on multiple platforms going forward.

The Scrypt implementation is also now Crypto++.